+++ sdk/communication/Azure.Communication.Common/CHANGELOG.md
@@ -5,6 +5,9 @@
 - Renamed `CommunicationUserCredential` to `CommunicationTokenCredential`.
 - Replace `CommunicationTokenCredential(bool refreshProactively, Func<CancellationToken, string> tokenRefresher,Func<CancellationToken, ValueTask<string>>? asyncTokenRefresher = null, string? initialToken = null)` 
 with `CommunicationTokenCredential(CommunicationTokenRefreshOptions tokenRefreshOptions)`.
+- Renamed `PhoneNumber` to `PhoneNumberIdentifier`
+- Renamed `CommunicationUser` to `CommunicationUserIdentifier `
+- Renamed `CallingApplication` to `CallingApplicationIdentifier`
 
 ## 1.0.0-beta.3 (2020-11-16)
 Updated `Azure.Communication.Common` version.